+++ manager/aica_django/converters/AICAStix.py
@@ -1,3 +1,4 @@
+from collections import OrderedDict
 import hashlib
 import json
 import uuid
@@ -15,7 +16,8 @@ from stix2 import (  # type:ignore
     NetworkTraffic,
     Note,
 )
-from stix2.base import _STIXBase  # type: ignore
+from stix2.base import _Extension, _STIXBase  # type: ignore
+from stix2.properties import StringProperty  # type: ignore
 from typing import Any, Dict, List, Optional, Union
 
 # These overwrite the default classes to create a deterministic ID where STIX doesn't otherwise provide one
@@ -255,3 +257,28 @@ class AICANote(Note):  # type: ignore
             content=content,
             authors=authors,
         )
+
+
+class DNP3RequestExt(_Extension):  # type: ignore
+    """For more detailed information on this object's properties, see
+    `the STIX 2.0 specification <http://docs.oasis-open.org/cti/stix/v2.0/cs01/part4-cyber-observable-objects/stix-v2.0-cs01-part4-cyber-observable-objects.html#_Toc496716262>`__.
+    """  # noqa
+
+    _type = "dnp3-request-ext"
+    _properties = OrderedDict(
+        [
+            ("dnp3_datalink_src", StringProperty(required=True)),
+            ("dnp3_datalink_dst", StringProperty(required=True)),
+            ("dnp3_application_unsolicited_from_slave", StringProperty(required=False)),
+            ("dnp3_datalink_from_master", StringProperty(required=True)),
+            ("dnp3_datalink_from_primary", StringProperty(required=True)),
+            ("dnp3_datalink_function", StringProperty(required=True)),
+            ("dnp3_application_function", StringProperty(required=False)),
+            ("dnp3_application_iin", StringProperty(required=False)),
+            ("dnp3_application_obj", StringProperty(required=False)),
+            ("dnp3_application_objq_code", StringProperty(required=False)),
+            ("dnp3_application_objq_index", StringProperty(required=False)),
+            ("dnp3_application_objq_prefix", StringProperty(required=False)),
+            ("dnp3_application_objq_range", StringProperty(required=False)),
+        ]
+    )
